+++ elements/wifi/station/associationrequester.cc
@@ -373,7 +373,7 @@ AssociationRequester_write_param(const String &in_s, Element *e, void *vparam,
 {
   AssociationRequester *f = (AssociationRequester *)e;
   String s = cp_uncomment(in_s);
-  switch((int)vparam) {
+  switch((intptr_t)vparam) {
   case H_DEBUG: {    //debug
     bool debug;
     if (!cp_bool(s, &debug))
